requirements.txt
numpy==2.3.3
Pillow==11.3.0
PyQt6==6.9.1
pyqt6_sip==13.10.2
scipy==1.16.1
pytest==8.4.2
pytest-qt==4.5.0